+++ src/lib/converter/factories/signature.ts
@@ -17,7 +17,6 @@ import type { Context } from "../context.js";
 import { ConverterEvents } from "../converter-events.js";
 import { convertDefaultValue } from "../convert-expression.js";
 import { removeUndefined } from "../utils/reflections.js";
-import { createSymbolId } from "./symbol-id.js";
 
 export function createSignature(
     context: Context,
@@ -51,7 +50,7 @@ export function createSignature(
     if (symbol && declaration) {
         context.project.registerSymbolId(
             sigRef,
-            createSymbolId(symbol, declaration),
+            context.createSymbolId(symbol, declaration),
         );
     }