+++ src/lib/converter/jsdoc.ts
@@ -14,7 +14,6 @@ import {
 import type { Context } from "./context.js";
 import { ConverterEvents } from "./converter-events.js";
 import { convertParameterNodes, convertTemplateParameterNodes } from "./factories/signature.js";
-import { createSymbolId } from "./factories/symbol-id.js";
 
 export function convertJsDocAlias(
     context: Context,
@@ -134,7 +133,7 @@ function convertJsDocSignature(context: Context, node: ts.JSDocSignature) {
     );
     context.project.registerSymbolId(
         signature,
-        createSymbolId(symbol, node),
+        context.createSymbolId(symbol, node),
     );
     context.registerReflection(signature, void 0);
     const signatureCtx = rc.withScope(signature);