+++ CHANGELOG.md
@@ -8,6 +8,7 @@ title: Changelog
 
 - If using JS config files, the `plugin` function can now be given plugin functions to load.
 - Permit `-` within tag names to support `typescript-json-schema`'s `@TJS-type` tag, #2972.
+- Exposed `Context.createSymbolId` for use by plugins.
 
 ### Thanks!